requirements.txt
selenium
deep-translator
python-telegram-bot==20.3
